update category module

+++ src/main/java/com/ratana/jfx/service/CategoryService.java
@@ -6,5 +6,5 @@ import java.util.List;
 
 public interface CategoryService {
     List<Category> findAll();
-    Category findById(int id);
+    Category findById(long id);
 }
